% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesquisadeprecos.R
\name{consultar_precos}
\alias{consultar_precos}
\title{Projeto IndigenR
Autor: Pedro Henrique Medeiros dos Santos (Funai/IPEA)
Consultar preços no Compras.gov}
\usage{
consultar_precos(
  codigo_item_catalogo,
  estado = NULL,
  servico = FALSE,
  regiao = NULL,
  meses = 12
)
}
\arguments{
\item{codigo_item_catalogo}{Código do item no catálogo}

\item{estado}{Sigla do estado (opcional)}

\item{servico}{Se TRUE, consulta serviços}

\item{regiao}{Região geográfica (opcional)}

\item{meses}{Quantidade de meses para filtro}
}
\value{
data.frame com resultados
}
\description{
Projeto IndigenR
Autor: Pedro Henrique Medeiros dos Santos (Funai/IPEA)
Consultar preços no Compras.gov
}
