% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_unidadeF.R
\name{padroniza_unidadeF}
\alias{padroniza_unidadeF}
\title{Padroniza unidades de fornecimento e medida}
\usage{
padroniza_unidadeF(dadosBrutos, servico = FALSE)
}
\arguments{
\item{dadosBrutos}{Data frame com os dados brutos da API}

\item{servico}{Se TRUE, indica que os dados são de serviços (default FALSE)}
}
\value{
Data frame com colunas padronizadas:
\itemize{
  \item unidadeFornecimento
  \item demais colunas originais
}
}
\description{
Esta função adapta os nomes e siglas das unidades de fornecimento,
garantindo consistência entre materiais e serviços.
}
\keyword{internal}
