% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_arquivo.R
\name{prepara_arquivo}
\alias{prepara_arquivo}
\title{Prepara arquivo Excel de resultados de pesquisa de preços}
\usage{
prepara_arquivo(dadosPrecos, nome_do_responsavel = "Nome", pasta)
}
\arguments{
\item{dadosPrecos}{Data frame com resultados saneados (deve conter coluna 'Status')}

\item{nome_do_responsavel}{Nome do responsável pela pesquisa}

\item{pasta}{Diretório onde o arquivo será salvo}
}
\value{
Caminho completo do arquivo Excel gerado
}
\description{
Prepara arquivo Excel de resultados de pesquisa de preços
}
