% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-compras.R
\name{puxa_api_compras_completo}
\alias{puxa_api_compras_completo}
\title{Extrai todos os dados paginados da API de Compras usando lista de parâmetros}
\usage{
puxa_api_compras_completo(
  url_base,
  parametros = list(),
  verbose = TRUE,
  max_tries = 5
)
}
\arguments{
\item{url_base}{\code{character}. A URL base do endpoint da API.}

\item{parametros}{\code{list}. Lista nomeada contendo os parâmetros da requisição.
(ex: \code{list(orgao = "00059311000126", codigoModalidade = 1:13)}). Suporta
no máximo um parâmetro com múltiplos valores (vetor) por vez.}

\item{verbose}{\code{logical}. Se \code{TRUE}, imprime mensagens de
progresso no console. Padrão: \code{TRUE}.}

\item{max_tries}{\code{integer}. Número máximo de tentativas de requisição
em caso de erro. Padrão: 5.}
}
\value{
Um \code{data.frame} com todos os registros combinados.
}
\description{
Utiliza a função \code{puxa_api_compras()} para iterar sobre todas as páginas de
um endpoint da API de Compras. A URL final é construída dinamicamente a partir
de uma lista de parâmetros fornecida pelo usuário. Caso um dos parâmetros seja
um vetor (ex: múltiplas modalidades), a função itera automaticamente sobre ele.
}
\examples{
\dontrun{
meus_filtros <- list(
  orgaoEntidadeCnpj = "00059311000126",
  codigoModalidade = 1:13 # Iteração automática habilitada!
)
dados <- puxa_api_compras_completo(
  url_base = "https://dadosabertos.compras.gov.br/modulo-contratacoes/1_consultarContratacoes_PNCP_14133",
  parametros = meus_filtros
)
}
}
