% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-pncp.R
\name{puxa_api_pncp_completo}
\alias{puxa_api_pncp_completo}
\title{Extrai todos os dados paginados da API PNCP}
\usage{
puxa_api_pncp_completo(
  url_base,
  parametros = list(),
  verbose = TRUE,
  max_tries = 5
)
}
\arguments{
\item{url_base}{URL com placeholders compatíveis com glue}

\item{parametros}{lista nomeada de parâmetros}

\item{verbose}{lógico}

\item{max_tries}{máximo de tentativas}
}
\value{
data.frame
}
\description{
Extrai todos os dados paginados da API PNCP
}
