% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siorg_estrutura_completa.R
\name{siorg_estrutura_completa}
\alias{siorg_estrutura_completa}
\title{Consulta completa de unidades organizacionais (UORGs)}
\usage{
siorg_estrutura_completa(codigo)
}
\arguments{
\item{codigo}{Código da unidade organizacional a ser consultada.}
}
\value{
Um \code{tibble} com os dados da UORG, incluindo informações de contato, endereço,
atos normativos e classificações. Caso a consulta não retorne resultados, retorna \code{NULL}.
}
\description{
Esta função consulta a base de Estrutura Organizacional do Governo Federal e retorna
informações detalhadas de uma unidade organizacional (UORG) a partir do seu código.
}
\examples{
\dontrun{
consulta_uorgs_completo("12345")
}
}
