% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodifica.R
\name{geocodifica_enderecos}
\alias{geocodifica_enderecos}
\title{Geocodifica endereços de unidades a partir de um relatório SIORG}
\usage{
geocodifica_enderecos(arq_siorg)
}
\arguments{
\item{arq_siorg}{Data frame contendo os dados de estrutura organizacional do SIORG.
Deve conter as colunas: \code{codigoUnidade}, \code{logradouro}, \code{numero},
\code{cep}, \code{bairro}, \code{nm_mun} e \code{uf}.}
}
\value{
Um objeto espacial do tipo \code{sf} contendo os dados das unidades
  geocodificadas e suas respectivas coordenadas.
}
\description{
Prepara os campos de endereço de um relatório de estrutura organizacional (qualquer
órgão) e realiza a geocodificação utilizando o pacote \pkg{geocodebr}.
}
\details{
A função realiza as seguintes etapas:
\itemize{
  \item remove duplicidades pelo código da unidade;
  \item padroniza os campos de endereço;
  \item geocodifica, resolvendo possíveis empates.
}
}
\examples{
\dontrun{
dados <- funai_relatorio_estrutura()
plot(sf::st_geometry(dados))
}

}
\seealso{
\code{\link[geocodebr]{geocode}},
\code{\link[geocodebr]{definir_campos}}
}
